Every time I interact with git, something goes wrong
Due to some issues with git, this entire commit got deleted when I was trying to push it. I managed to recover almost everything and I redid the few things it missed, but there's a tiny chance that I missed something

Fixed lots of multiplayer stuff

Follow-Up now has a Super form variant that replaces the secondary skill after using IDW Attack. It's a very simple attack but it is very fun to be able to easily kick and launch very large enemies

Reworded some skill descriptions slightly

Sonic Boom's damage buff I was planning on giving it has been reverted. Instead, it now applies a new debuff on hit that reduces armor. This was based on a suggestion from astrasilverhorn on Discord

Added new buff icons for stuff related to Sonic's parry and the new Sonic Boom debuffs

Scepter Boost now launches up to mediumweight enemies like Follow Up can

Very minor known issue with Grand Slams new effect of holding enemies in place. It only works if the target is under authority of the host, so it won't work when used against client players. Very niche though so I don't know if I'm going to bother with rethinking things to fix it

Might be issues with Power Boost HUD not showing up consistently in multiplayer. I previously ran into issues of the HUD not working if you died and respawned as Sonic while playing with another Sonic. Didn't test if that got fixed so I want to look into that soon